Add guard clause around response
Most of the commands by the engine don't require a response, but are
instead just information that needs to be parsed. These commands still
triggered an output message, which was picked up by the game engine as
the response to the move request. Of course, the output was no valid
move, so the bot was disqualified after a few rounds.

To fix this, a guard clause is added that validated that there is
actually a message to send back. This does not prevent invalid responses
from being sent to the engine, but at least it provides a way to filter
the output.

It is part of the parser's implementations to make sure that they either
return something useful or nothing at all for any input by the game
engine.