Move logic from AbstractParser to Parser and delete it
The scope of this gem is just to provide a parser. Therefore, it
is not necessary to scope the implementation in its own class under
the Parser namespace. To make things less complex, the implementation
of the AbstractParser has been moved into the Parser, which has
been converted from a module to a class. Since we are still using
the AIGames namespace, we don't expect this to have any side effects
with other gems.